opentelemetry-http: Enhance OpenTelemetryHttpRequesterFilterTest
Motivation:

We want to make sure our request observers are able to see the
right spans and add attributes.

Modifications:

Reuse the TestHttpLifecycleObserver that we have built out for the
server side.

Result:

Better test coverage.
